Move external enrollment classes, tasks method and factory class to its own file.
