+++ openedx_external_enrollments/edxapp_wrapper/get_student.py
@@ -1,4 +1,4 @@
-"""Courseware definitions."""
+"""Student definitions."""
 
 from importlib import import_module
 from django.conf import settings
